Move some swath tests to test_geometry/test_swath.py
