+++ pyresample/test/test_geometry_legacy.py
@@ -17,7 +17,6 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """Test the geometry objects."""
-import contextlib
 import random
 import sys
 import unittest
@@ -29,184 +28,18 @@ import pytest
 import xarray as xr
 from pyproj import CRS, Proj
 
-from pyresample import geo_filter, geometry, parse_area_file
+from pyresample import geo_filter, geometry
 from pyresample.geometry import (
     IncompatibleAreas,
     SwathDefinition,
     combine_area_extents_vertical,
     concatenate_area_defs,
 )
-from pyresample.test.utils import (
-    catch_warnings,
-    create_test_latitude,
-    create_test_longitude,
-)
+from pyresample.test.utils import catch_warnings
 
 
 class Test(unittest.TestCase):
     """Unit testing the geometry and geo_filter modules."""
-
-    def test_dump(self):
-        """Test exporting area defs."""
-        from io import StringIO
-
-        import yaml
-
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)',
-                                           'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '90.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                            -909968.64000000001,
-                                            1029087.28,
-                                            1490031.3600000001])
-        res = yaml.safe_load(area_def.dump())
-        expected = yaml.safe_load(('areaD:\n  description: Europe (3km, HRV, VTC)\n'
-                                   '  projection:\n    a: 6378144.0\n    b: 6356759.0\n'
-                                   '    lat_0: 90.0\n    lat_ts: 50.0\n    lon_0: 8.0\n'
-                                   '    proj: stere\n  shape:\n    height: 800\n'
-                                   '    width: 800\n  area_extent:\n'
-                                   '    lower_left_xy: [-1370912.72, -909968.64]\n'
-                                   '    upper_right_xy: [1029087.28, 1490031.36]\n'))
-
-        self.assertEqual(set(res.keys()), set(expected.keys()))
-        res = res['areaD']
-        expected = expected['areaD']
-        self.assertEqual(set(res.keys()), set(expected.keys()))
-        self.assertEqual(res['description'], expected['description'])
-        self.assertEqual(res['shape'], expected['shape'])
-        self.assertEqual(res['area_extent']['lower_left_xy'],
-                         expected['area_extent']['lower_left_xy'])
-        # pyproj versions may effect how the PROJ is formatted
-        for proj_key in ['a', 'lat_0', 'lon_0', 'proj', 'lat_ts']:
-            self.assertEqual(res['projection'][proj_key],
-                             expected['projection'][proj_key])
-
-        # EPSG
-        projections = {
-            '+init=epsg:3006': 'init: epsg:3006',
-            'EPSG:3006': 'EPSG: 3006',
-        }
-
-        for projection, epsg_yaml in projections.items():
-            area_def = geometry.AreaDefinition('baws300_sweref99tm', 'BAWS, 300m resolution, sweref99tm',
-                                               'sweref99tm',
-                                               projection,
-                                               4667,
-                                               4667,
-                                               [-49739, 5954123, 1350361, 7354223])
-            res = yaml.safe_load(area_def.dump())
-            yaml_string = ('baws300_sweref99tm:\n'
-                           '  description: BAWS, 300m resolution, sweref99tm\n'
-                           '  projection:\n'
-                           '    {epsg}\n'
-                           '  shape:\n'
-                           '    height: 4667\n'
-                           '    width: 4667\n'
-                           '  area_extent:\n'
-                           '    lower_left_xy: [-49739, 5954123]\n'
-                           '    upper_right_xy: [1350361, 7354223]\n'.format(epsg=epsg_yaml))
-            expected = yaml.safe_load(yaml_string)
-        self.assertDictEqual(res, expected)
-
-        # testing writing to file with file-like object
-        sio = StringIO()
-        area_def.dump(filename=sio)
-        res = yaml.safe_load(sio.getvalue())
-        self.assertDictEqual(res, expected)
-
-        # test writing to file with string filename
-        with patch('pyresample.geometry.open') as mock_open:
-            area_def.dump(filename='area_file.yml')
-            mock_open.assert_called_once_with('area_file.yml', 'a')
-            mock_open.return_value.__enter__().write.assert_called_once_with(yaml_string)
-
-    def test_dump_numpy_extents(self):
-        """Test exporting area defs when extents are Numpy floats."""
-        import yaml
-
-        area_def = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)',
-                                           'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '90.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [np.float64(-1370912.72),
-                                            np.float64(-909968.64000000001),
-                                            np.float64(1029087.28),
-                                            np.float64(1490031.3600000001)])
-        res = yaml.safe_load(area_def.dump())
-        expected = yaml.safe_load(('areaD:\n  description: Europe (3km, HRV, VTC)\n'
-                                   '  projection:\n    a: 6378144.0\n    b: 6356759.0\n'
-                                   '    lat_0: 90.0\n    lat_ts: 50.0\n    lon_0: 8.0\n'
-                                   '    proj: stere\n  shape:\n    height: 800\n'
-                                   '    width: 800\n  area_extent:\n'
-                                   '    lower_left_xy: [-1370912.72, -909968.64]\n'
-                                   '    upper_right_xy: [1029087.28, 1490031.36]\n'))
-
-        self.assertEqual(res['areaD']['area_extent']['lower_left_xy'],
-                         expected['areaD']['area_extent']['lower_left_xy'])
-
-    def test_parse_area_file(self):
-        """Test parsing the are file."""
-        expected = geometry.AreaDefinition('areaD', 'Europe (3km, HRV, VTC)',
-                                           'areaD',
-                                           {'a': '6378144.0',
-                                            'b': '6356759.0',
-                                            'lat_0': '50.00',
-                                            'lat_ts': '50.00',
-                                            'lon_0': '8.00',
-                                            'proj': 'stere'},
-                                           800,
-                                           800,
-                                           [-1370912.72,
-                                            -909968.64000000001,
-                                            1029087.28,
-                                            1490031.3600000001])
-        yaml_str = ('areaD:\n  description: Europe (3km, HRV, VTC)\n'
-                    '  projection:\n    a: 6378144.0\n    b: 6356759.0\n'
-                    '    lat_0: 50.0\n    lat_ts: 50.0\n    lon_0: 8.0\n'
-                    '    proj: stere\n  shape:\n    height: 800\n'
-                    '    width: 800\n  area_extent:\n'
-                    '    lower_left_xy: [-1370912.72, -909968.64]\n'
-                    '    upper_right_xy: [1029087.28, 1490031.36]\n')
-        area_def = parse_area_file(yaml_str, 'areaD')[0]
-        self.assertEqual(area_def, expected)
-
-        # EPSG
-        projections = {
-            '+init=epsg:3006': 'init: epsg:3006',
-            'EPSG:3006': 'EPSG: 3006',
-        }
-        for projection, epsg_yaml in projections.items():
-            expected = geometry.AreaDefinition('baws300_sweref99tm', 'BAWS, 300m resolution, sweref99tm',
-                                               'sweref99tm',
-                                               projection,
-                                               4667,
-                                               4667,
-                                               [-49739, 5954123, 1350361, 7354223])
-            yaml_str = ('baws300_sweref99tm:\n'
-                        '  description: BAWS, 300m resolution, sweref99tm\n'
-                        '  projection:\n'
-                        '    {epsg}\n'
-                        '  shape:\n'
-                        '    height: 4667\n'
-                        '    width: 4667\n'
-                        '  area_extent:\n'
-                        '    lower_left_xy: [-49739, 5954123]\n'
-                        '    upper_right_xy: [1350361, 7354223]'.format(epsg=epsg_yaml))
-            area_def = parse_area_file(yaml_str, 'baws300_sweref99tm')[0]
-            self.assertEqual(area_def, expected)
 
     def test_base_type(self):
         """Test the base type."""
@@ -2934,119 +2767,6 @@ class TestAreaDefGetAreaSlices(unittest.TestCase):
         assert slice_cols == expected_slice_cols
 
 
-class TestBboxLonlats:
-    """Test 'get_bbox_lonlats' for various geometry cases."""
-
-    @pytest.mark.parametrize(
-        ("lon_start", "lon_stop", "lat_start", "lat_stop", "exp_nonforced_clockwise"),
-        [
-            (3.0, 12.0, 75.0, 26.0, True),  # [0, 0] at north-west corner
-            (12.0, 3.0, 75.0, 26.0, False),  # [0, 0] at north-east corner
-            (3.0, 12.0, 26.0, 75.0, False),  # [0, 0] at south-west corner
-            (12.0, 3.0, 26.0, 75.0, True),  # [0, 0] at south-east corner
-        ]
-    )
-    @pytest.mark.parametrize("force_clockwise", [False, True])
-    @pytest.mark.parametrize("use_dask", [False, True])
-    @pytest.mark.parametrize("use_xarray", [False, True])
-    @pytest.mark.parametrize("nan_pattern", [None, "scan", "half", "whole"])
-    def test_swath_def_bbox(self, lon_start, lon_stop,
-                            lat_start, lat_stop, exp_nonforced_clockwise,
-                            force_clockwise,
-                            use_dask, use_xarray, nan_pattern):
-        swath_shape = (50, 10)
-        lons = create_test_longitude(lon_start, lon_stop, swath_shape)
-        lats = create_test_latitude(lat_start, lat_stop, swath_shape)
-        lons, lats = _add_nans_if_necessary(lons, lats, nan_pattern)
-        lons, lats = _convert_type_if_necessary(lons, lats, use_dask, use_xarray)
-        swath_def = SwathDefinition(lons, lats)
-        with _raises_if(nan_pattern == "whole", ValueError):
-            bbox_lons, bbox_lats = swath_def.get_bbox_lonlats(force_clockwise=force_clockwise)
-        if nan_pattern != "whole":
-            _check_bbox_structure_and_values(bbox_lons, bbox_lats)
-            _check_bbox_clockwise(bbox_lons, bbox_lats, exp_nonforced_clockwise, force_clockwise)
-
-    def test_swath_def_bbox_decimated(self):
-        from pyresample.geometry import SwathDefinition
-
-        lons, lats = _gen_swath_lons_lats()
-        swath_def = SwathDefinition(lons, lats)
-        bbox_lons, bbox_lats = swath_def.get_bbox_lonlats(frequency=None)
-        assert len(bbox_lons) == len(bbox_lats)
-        assert len(bbox_lons) == 4
-        assert len(bbox_lons[0]) == 10
-        assert len(bbox_lons[1]) == 50
-        assert len(bbox_lons[2]) == 10
-        assert len(bbox_lons[3]) == 50
-
-        bbox_lons, bbox_lats = swath_def.get_bbox_lonlats(frequency=5)
-        assert len(bbox_lons) == len(bbox_lats)
-        assert len(bbox_lons) == 4
-        assert len(bbox_lons[0]) == 5
-        assert len(bbox_lons[1]) == 5
-        assert len(bbox_lons[2]) == 5
-        assert len(bbox_lons[3]) == 5
-        assert bbox_lons[0][-1] == bbox_lons[1][0]
-
-
-def _add_nans_if_necessary(lons, lats, nan_pattern):
-    if nan_pattern == "scan":
-        lons[20:30, -1] = np.nan
-    elif nan_pattern == "half":
-        lons[:25, -1] = np.nan
-    elif nan_pattern == "whole":
-        lons[:, -1] = np.nan
-    return lons, lats
-
-
-def _convert_type_if_necessary(lons, lats, use_dask, use_xarray):
-    if use_dask:
-        lons = da.from_array(lons)
-        lats = da.from_array(lats)
-    if use_xarray:
-        lons = xr.DataArray(lons, dims=('y', 'x'))
-        lats = xr.DataArray(lats, dims=('y', 'x'))
-    return lons, lats
-
-
-def _check_bbox_structure_and_values(bbox_lons, bbox_lats):
-    assert not any(np.isnan(side_lon).any() for side_lon in bbox_lons)
-    assert not any(np.isnan(side_lat).any() for side_lat in bbox_lats)
-    assert len(bbox_lons) == len(bbox_lats)
-    assert len(bbox_lons) == 4
-    for side_lons, side_lats in zip(bbox_lons, bbox_lats):
-        assert isinstance(side_lons, np.ndarray)
-        assert isinstance(side_lats, np.ndarray)
-        assert side_lons.shape == side_lats.shape
-
-
-def _check_bbox_clockwise(bbox_lons, bbox_lats, exp_nonforced_clockwise, force_clockwise):
-    is_cw = _is_clockwise(np.concatenate(bbox_lons), np.concatenate(bbox_lats))
-    if exp_nonforced_clockwise or force_clockwise:
-        assert is_cw
-    else:
-        assert not is_cw
-
-
-@contextlib.contextmanager
-def _raises_if(condition, exp_exception, *args, **kwargs):
-    expectation = pytest.raises(exp_exception, *args, **kwargs) if condition else contextlib.nullcontext()
-    with expectation:
-        yield
-
-
-def _is_clockwise(lons, lats):
-    # https://stackoverflow.com/a/1165943/433202
-    prev_point = (lons[0], lats[0])
-    edge_sum = 0
-    for point in zip(lons[1:], lats[1:]):
-        xdiff = point[0] - prev_point[0]
-        ysum = point[1] + prev_point[1]
-        edge_sum += xdiff * ysum
-        prev_point = point
-    return edge_sum > 0
-
-
 class TestBoundary(unittest.TestCase):
     """Test 'boundary' method for <area_type>Definition classes."""
 
@@ -3276,68 +2996,3 @@ class TestBoundary(unittest.TestCase):
                                       [1.3, 65.86],
                                       [1.2, 65.89]])
         assert np.allclose(expected_vertices, boundary.vertices)
-
-
-def _gen_swath_def_xarray_dask():
-    lons, lats = _gen_swath_lons_lats()
-    lons_dask = da.from_array(lons)
-    lats_dask = da.from_array(lats)
-    lons_xr = xr.DataArray(lons_dask, dims=('y', 'x'))
-    lats_xr = xr.DataArray(lats_dask, dims=('y', 'x'))
-    return SwathDefinition(lons_xr, lats_xr)
-
-
-def _gen_swath_def_xarray_numpy():
-    lons, lats = _gen_swath_lons_lats()
-    lons_xr = xr.DataArray(lons, dims=('y', 'x'))
-    lats_xr = xr.DataArray(lats, dims=('y', 'x'))
-    return SwathDefinition(lons_xr, lats_xr)
-
-
-def _gen_swath_def_dask():
-    lons, lats = _gen_swath_lons_lats()
-    lons_dask = da.from_array(lons)
-    lats_dask = da.from_array(lats)
-    return SwathDefinition(lons_dask, lats_dask)
-
-
-def _gen_swath_def_numpy():
-    lons, lats = _gen_swath_lons_lats()
-    return SwathDefinition(lons, lats)
-
-
-def _gen_swath_lons_lats():
-    swath_shape = (50, 10)
-    lon_start, lon_stop, lat_start, lat_stop = (3.0, 12.0, 75.0, 26.0)
-    lons = create_test_longitude(lon_start, lon_stop, swath_shape)
-    lats = create_test_latitude(lat_start, lat_stop, swath_shape)
-    return lons, lats
-
-
-class TestGeometryHashability:
-    """Test geometry objects being hashable and other related uses."""
-
-    @pytest.mark.parametrize(
-        "swath_def_func1",
-        [
-            _gen_swath_def_numpy,
-            _gen_swath_def_dask,
-            _gen_swath_def_xarray_numpy,
-            _gen_swath_def_xarray_dask,
-        ])
-    def test_swath_as_dict_keys(self, swath_def_func1):
-        from .utils import assert_maximum_dask_computes
-        swath_def1 = swath_def_func1()
-        swath_def2 = swath_def_func1()
-
-        with assert_maximum_dask_computes(0):
-            assert hash(swath_def1) == hash(swath_def2)
-
-            test_dict = {}
-            test_dict[swath_def1] = 5
-            assert test_dict[swath_def1] == 5
-            assert test_dict[swath_def2] == 5
-            assert test_dict.get(swath_def2) == 5
-            test_dict[swath_def2] = 6
-            assert test_dict[swath_def1] == 6
-            assert test_dict[swath_def2] == 6
